refractor client not namespaced to client (namespaced client) and recovery (no namespace)
