+++ models/recovery/index.js
@@ -1,0 +1,8 @@
+module.exports = globalConfig => ({
+  namespace: null,
+  effects: require('./effects')(globalConfig),
+  reducers: require('./reducers')(globalConfig),
+  state: {
+    recoveryPossible: false
+  }
+})
