+++ pages/choseUsername.js
@@ -4,7 +4,7 @@ const html = require('choo/html')
 
 module.exports = function (globalConfig) {
   return function (state, prev, send) {
-    if (state.connectivityState === globalConfig.connectivityStates.none) {
+    if (state.client.connectivityState === globalConfig.connectivityStates.none) {
       send('location:set', '/')
       return html`<div></div>`
     }
@@ -17,7 +17,7 @@ module.exports = function (globalConfig) {
           <h1>Enter a Name</h1>
       </div>
       <div class="row">
-          <input type="text" id="username" name="username" class="enter_id" value="${state.username ? state.username : ''}" autofocus="autofocus" >
+          <input type="text" id="username" name="username" class="enter_id" value="${state.client.username ? state.client.username : ''}" autofocus="autofocus" >
           <button class="login_start" onclick=${publishName}>Confirm</button>
       </div>
     </div>
@@ -29,7 +29,7 @@ module.exports = function (globalConfig) {
       event.preventDefault()
       var username = document.getElementById('username').value
       if (username.length > 0) {
-        send('setUsername', username)
+        send('client:setUsername', username)
         send('location:set', '/game')
       }
     }
