+++ pages/connecting.js
@@ -4,12 +4,12 @@ const html = require('choo/html')
 
 module.exports = function (globalConfig) {
   return function (state, prev, send) {
-    if (state.connectivityState === globalConfig.connectivityStates.none) {
+    if (state.client.connectivityState === globalConfig.connectivityStates.none) {
       send('location:set', '/')
     }
 
-    if (state.connectivityState === globalConfig.connectivityStates.connected) {
-      if (!state.username) {
+    if (state.client.connectivityState === globalConfig.connectivityStates.connected) {
+      if (!state.client.username) {
         send('location:set', '/choseUsername')
       } else {
         send('location:set', '/game')
