+++ pages/recovery.js
@@ -2,11 +2,11 @@ const html = require('choo/html')
 
 module.exports = function (globalConfig) {
   return function (state, prev, send) {
-    if (state.connectivityState === globalConfig.connectivityStates.none) {
+    if (state.client.connectivityState === globalConfig.connectivityStates.none) {
       send('location:set', '/')
     }
 
-    if (state.connectivityState === globalConfig.connectivityStates.connected) {
+    if (state.client.connectivityState === globalConfig.connectivityStates.connected) {
       send('location:set', '/choseUsername')
     }
